add tooltip to timer buttons, minor adjustment of timer error prompt
