+++ src/components/Timer/Countdown.module.css
@@ -95,7 +95,7 @@
   display: flex;
   justify-content: space-between;
   align-items: center;
-  margin-top: 0.8rem;
+  margin-top: 1rem;
   gap: 0.5rem;
 }
 
@@ -206,4 +206,10 @@
   text-align: center;
   font-weight: bold;
   font-size: 1.5rem;
+}
+
+.transitionColor {
+  transition-property: color, background-color, border-color, text-decoration-color, fill, stroke;
+  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+  transition-duration: 300ms;
 }
