+++ src/components/Timer/Timer.module.css
@@ -2,6 +2,7 @@
   position: relative;
   display: flex;
   align-items: center;
+  padding-right: 1rem;
   gap: 0.5rem;
   z-index: 30;
   user-select: none;
@@ -102,8 +103,4 @@
   transition-property: color, background-color, border-color, text-decoration-color, fill, stroke;
   transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
   transition-duration: 300ms;
-}
-
-.btnStop:disabled {
-  cursor: not-allowed;
 }
